moving ValidatorExtensions into root namespace and deleting ExtensionMethods folder from project.
