+++ Validator/Validator.csproj
@@ -41,7 +41,6 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ExtensionMethods\ValidatorExtensions.cs" />
     <Compile Include="IpVersion.cs" />
     <Compile Include="IsBase64.cs" />
     <Compile Include="IsbnVersion.cs" />
@@ -55,7 +54,9 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="UuidVersion.cs" />
     <Compile Include="Validator.cs" />
+    <Compile Include="ValidatorExtensions.cs" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
